Reorganizing Parameter and DataColleciton Objects
