+++ tests/pmv_test.py
@@ -3,7 +3,7 @@
 import unittest
 import pytest
 
-from ladybug.comfort.standard.pmv import PMVParameters
+from ladybug.comfort.parameter.pmv import PMVParameter
 
 from ladybug.comfort.pmv import pmv, fanger_pmv, pierce_set, ppd_from_pmv, \
     pmv_from_ppd, calc_missing_pmv_input
@@ -92,14 +92,14 @@ class PMVTestCase(unittest.TestCase):
         assert updated_input_6['clo'] == pytest.approx(0.6546, rel=1e-2)
         assert updated_input_7['wme'] == pytest.approx(0.3577, rel=1e-2)
 
-    def test_pmv_parameters(self):
-        """Test PMVParameters."""
+    def test_pmv_parameter(self):
+        """Test PMVParameter."""
         ppd_comfort_thresh = 20
         humid_ratio_up = 0.012
         humid_ratio_low = 0.004
         still_air_thresh = 0.2
 
-        pmv_comf = PMVParameters(
+        pmv_comf = PMVParameter(
             ppd_comfort_thresh, humid_ratio_up, humid_ratio_low, still_air_thresh)
 
         assert pmv_comf.ppd_comfort_thresh == ppd_comfort_thresh
@@ -107,41 +107,41 @@ class PMVTestCase(unittest.TestCase):
         assert pmv_comf.humid_ratio_lower == humid_ratio_low
         assert pmv_comf.still_air_threshold == still_air_thresh
 
-    def test_pmv_parameters_invalid(self):
-        """Test PMVParameters for invalid inputs."""
+    def test_pmv_parameter_invalid(self):
+        """Test PMVParameter for invalid inputs."""
         ppd_comfort_thresh = 110
         humid_ratio_up = 12
         humid_ratio_low = -1
         still_air_thresh = -1
 
         with pytest.raises(AssertionError):
-            PMVParameters(ppd_comfort_thresh=ppd_comfort_thresh)
+            PMVParameter(ppd_comfort_thresh=ppd_comfort_thresh)
         with pytest.raises(AssertionError):
-            PMVParameters(humid_ratio_upper=humid_ratio_up)
+            PMVParameter(humid_ratio_upper=humid_ratio_up)
         with pytest.raises(AssertionError):
-            PMVParameters(humid_ratio_lower=humid_ratio_low)
+            PMVParameter(humid_ratio_lower=humid_ratio_low)
         with pytest.raises(AssertionError):
-            PMVParameters(still_air_threshold=still_air_thresh)
+            PMVParameter(still_air_threshold=still_air_thresh)
 
     def test_comfort_check(self):
-        """Test comfort check on PMVParameters."""
-        pmv_comf = PMVParameters()
+        """Test comfort check on PMVParameter."""
+        pmv_comf = PMVParameter()
         comf_test = pmv_comf.is_comfortable(13, 0.01)
         assert comf_test is False
         comf_test = pmv_comf.is_comfortable(7, 0.01)
         assert comf_test is True
 
     def test_thermal_condition_check(self):
-        """Test the thermal condition check on PMVParameters."""
-        pmv_comf = PMVParameters()
+        """Test the thermal condition check on PMVParameter."""
+        pmv_comf = PMVParameter()
         condition_test = pmv_comf.thermal_condition(-1, 20)
         assert condition_test == -1
         condition_test = pmv_comf.thermal_condition(0, 5)
         assert condition_test == 0
 
     def test_discomfort_reason_check(self):
-        """Test the thermal condition check on PMVParameters."""
-        pmv_comf = PMVParameters()
+        """Test the thermal condition check on PMVParameter."""
+        pmv_comf = PMVParameter()
         condition_test = pmv_comf.discomfort_reason(-1, 20, 0.01)
         assert condition_test == -1
         condition_test = pmv_comf.discomfort_reason(0, 5, 0.01)
